+++ modulemate/src/main/java/at/xa1/modulemate/mode/flash/FlashMode.kt
@@ -1,8 +1,8 @@
-package at.xa1.modulemate.flashmode
+package at.xa1.modulemate.mode.flash
 
 import at.xa1.modulemate.command.Command
 import at.xa1.modulemate.command.CommandList
-import at.xa1.modulemate.liveui.LiveUiMode
+import at.xa1.modulemate.mode.LiveUiMode
 import at.xa1.modulemate.ui.ListBox
 import at.xa1.modulemate.ui.ListBoxItemRenderer
 import at.xa1.modulemate.ui.TextBox
