+++ modulemate/src/main/java/at/xa1/modulemate/mode/flash/FlashScreen.kt
@@ -1,6 +1,7 @@
-package at.xa1.modulemate.helpmode
+package at.xa1.modulemate.mode.flash
 
 import at.xa1.modulemate.cli.CliColor
+import at.xa1.modulemate.command.Command
 import at.xa1.modulemate.ui.ListBox
 import at.xa1.modulemate.ui.ScreenContext
 import at.xa1.modulemate.ui.TextBox
@@ -8,10 +9,9 @@ import at.xa1.modulemate.ui.UiUserInput
 import at.xa1.modulemate.ui.moveDown
 import at.xa1.modulemate.ui.moveUp
 import at.xa1.modulemate.ui.print
-import at.xa1.modulemate.ui.reduce
 import at.xa1.modulemate.ui.updateHeight
 
-fun helpScreen(context: ScreenContext, state: HelpScreenState) = context.printScreen {
+fun flashScreen(context: ScreenContext, state: FlashScreenState) = context.printScreen {
     print(state.searchBox, context.size.columns)
 
     print(state.listBox)
@@ -20,25 +20,17 @@ fun helpScreen(context: ScreenContext, state: HelpScreenState) = context.printSc
     print(CliColor.cursorDown(1) + CliColor.cursorRight(5 + state.searchBox.cursor))
 }
 
-data class HelpScreenState(
+data class FlashScreenState(
     val searchBox: TextBox,
-    val listBox: ListBox<String>
+    val listBox: ListBox<Command>
 )
 
-fun HelpScreenState.reduce(input: UiUserInput, height: Int): HelpScreenState {
+fun FlashScreenState.reduce(input: UiUserInput, height: Int): FlashScreenState {
     return when (input) {
         UiUserInput.Arrow.Down -> copy(listBox = listBox.moveDown())
         UiUserInput.Arrow.Up -> copy(listBox = listBox.moveUp())
-
-        UiUserInput.Return -> {
-            println("TODO execute")
-            this
-        }
-        else -> {
-            copy(searchBox = searchBox.reduce(input))
-        }
+        else -> this
     }.updateHeight(height)
 }
-
-internal fun HelpScreenState.updateHeight(height: Int): HelpScreenState =
+internal fun FlashScreenState.updateHeight(height: Int): FlashScreenState =
     copy(listBox = listBox.updateHeight(height - 4))
