+++ modulemate/src/main/java/at/xa1/modulemate/mode/help/HelpScreen.kt
@@ -1,7 +1,6 @@
-package at.xa1.modulemate.flashmode
+package at.xa1.modulemate.mode.help
 
 import at.xa1.modulemate.cli.CliColor
-import at.xa1.modulemate.command.Command
 import at.xa1.modulemate.ui.ListBox
 import at.xa1.modulemate.ui.ScreenContext
 import at.xa1.modulemate.ui.TextBox
@@ -9,9 +8,10 @@ import at.xa1.modulemate.ui.UiUserInput
 import at.xa1.modulemate.ui.moveDown
 import at.xa1.modulemate.ui.moveUp
 import at.xa1.modulemate.ui.print
+import at.xa1.modulemate.ui.reduce
 import at.xa1.modulemate.ui.updateHeight
 
-fun flashScreen(context: ScreenContext, state: FlashScreenState) = context.printScreen {
+fun helpScreen(context: ScreenContext, state: HelpScreenState) = context.printScreen {
     print(state.searchBox, context.size.columns)
 
     print(state.listBox)
@@ -20,17 +20,25 @@ fun flashScreen(context: ScreenContext, state: FlashScreenState) = context.print
     print(CliColor.cursorDown(1) + CliColor.cursorRight(5 + state.searchBox.cursor))
 }
 
-data class FlashScreenState(
+data class HelpScreenState(
     val searchBox: TextBox,
-    val listBox: ListBox<Command>
+    val listBox: ListBox<String>
 )
 
-fun FlashScreenState.reduce(input: UiUserInput, height: Int): FlashScreenState {
+fun HelpScreenState.reduce(input: UiUserInput, height: Int): HelpScreenState {
     return when (input) {
         UiUserInput.Arrow.Down -> copy(listBox = listBox.moveDown())
         UiUserInput.Arrow.Up -> copy(listBox = listBox.moveUp())
-        else -> this
+
+        UiUserInput.Return -> {
+            println("TODO execute")
+            this
+        }
+        else -> {
+            copy(searchBox = searchBox.reduce(input))
+        }
     }.updateHeight(height)
 }
-internal fun FlashScreenState.updateHeight(height: Int): FlashScreenState =
+
+internal fun HelpScreenState.updateHeight(height: Int): HelpScreenState =
     copy(listBox = listBox.updateHeight(height - 4))
