+++ modulemate/src/main/java/at/xa1/modulemate/mode/modules/ModulesMode.kt
@@ -1,24 +1,24 @@
-package at.xa1.modulemate.searchmode
+package at.xa1.modulemate.mode.modules
 
-import at.xa1.modulemate.liveui.LiveUiMode
-import at.xa1.modulemate.liveui.ModulesListItemRenderer
+import at.xa1.modulemate.mode.LiveUiMode
+import at.xa1.modulemate.mode.ModulesListItemRenderer
 import at.xa1.modulemate.module.Modules
 import at.xa1.modulemate.ui.ListBox
 import at.xa1.modulemate.ui.TextBox
 import at.xa1.modulemate.ui.Ui
 import at.xa1.modulemate.ui.UiUserInput
 
-class SearchMode(
+class ModulesMode(
     private val ui: Ui,
     private val modules: Modules
 ) : LiveUiMode {
-    private var state = SearchScreenState(
+    private var state = ModulesScreenState(
         searchBox = TextBox(
-            hint = "Search Mode: Type to search modules and commands",
-            emoji = "🔍"
+            hint = "Module Mode",
+            emoji = "📦"
         ),
         listBox = ListBox(
-            items = modules.allModules,
+            items = modules.recentModules,
             height = 0,
             renderer = ModulesListItemRenderer(modules)
         )
@@ -27,14 +27,13 @@ class SearchMode(
     override fun print(input: UiUserInput?) {
         val context = ui.createScreenContext()
 
-        state = if (input != null) {
-            state.reduce(input, modules, context.size.rows)
-        } else {
-            state.updateHeight(context.size.rows)
+        if (input != null) {
+            state = state.reduce(input, modules, context.size.rows)
         }
+        state = state.updateList(modules).updateHeight(context.size.rows)
 
         context.printScreen {
-            searchScreen(this, state)
+            modulesScreen(this, state)
         }
     }
 }
